requirements.txt
gitpython==3.1.30
openai==0.28.0
psycopg2==2.9.7
tiktoken==0.2.0
pydantic==1.10.2
requests==2.31.0
pytest==7.4.2 
